backend/requirements.txt
# Backend Dependencies — BioAI Hub

# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1

# Rate Limiting
django-ratelimit==4.1.0

# Utils
python-decouple==3.8
python-dotenv==1.0.0
dj-database-url==2.1.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-bdd==6.1.1
factory-boy==3.3.0
faker==20.1.0
freezegun==1.4.0

# Code Quality
black==23.12.1
flake8==6.1.0
isort==5.13.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Observability
django-extensions==3.2.3
